Rename SyntaxTreeContextExtensions
